requirements.txt
Django==5.0.2
django-compat==1.0.15
django-cors-headers==4.7.0
djangorestframework==3.16.1
django-filter>=24,<26
dj-database-url==2.0.0
pymysql
django-rest-auth
python-dateutil==2.9.0.post0
requests==2.32.3
cryptography==42.0.8
google-cloud-storage==2.15.0
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
mysqlclient==2.2.0  # Si sigues usando MySQL
# O para PostgreSQL:
# psycopg2-binary==2.9.9